feat: abstraction for greater test coverage
