+++ src/Configuration/Provisioning/DashboardDatatable/index.js
@@ -1,0 +1,3 @@
+import DashboardDatatable from './DashboardDatatable';
+
+export default DashboardDatatable;
